chore: refactor download multiple files func
