+++ src/renderer/features/wallets/ExportKeys/lib/download-multiple-files.ts
@@ -1,26 +1,29 @@
-type FileForDownload = { url: string; fileName: string };
+type FileForDownload = { blob: Blob; fileName: string };
 
 export const downloadFiles = (files: FileForDownload[]) => {
+  const a = document.createElement('a');
+
   function downloadNext(index: number) {
     if (index >= files.length) {
       return;
     }
-    const a = document.createElement('a');
-    a.href = files[index].url;
-    a.target = '_parent';
-    if ('download' in a) {
-      a.download = files[index].fileName;
-    }
 
-    (document.body || document.documentElement).appendChild(a);
+    const file = files[index];
+    const url = window.URL.createObjectURL(file.blob);
+
+    document.body.appendChild(a);
+
+    a.href = url;
+    a.download = file.fileName;
     a.click();
 
-    a.parentNode?.removeChild(a);
-    // Download the next file with a small timeout. The timeout is necessary
-    // for IE, which will otherwise only download the first file.
-    setTimeout(function () {
-      downloadNext(index + 1);
-    }, 500);
+    setTimeout(() => {
+      window.URL.revokeObjectURL(url);
+    }, 0);
+
+    downloadNext(index + 1);
   }
   downloadNext(0);
+
+  document.body.removeChild(a);
 };
